requirements.txt
Django==2.2.4
pytz==2019.2
sqlparse==0.3.0
requests==2.22.0
djangorestframework==3.10.2
markdown==3.1.1
django-filter==2.2.0